// web/src/api/generated/models/chapter-summary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ChapterId } from './chapter-id';
// May contain unused imports in some cases
// @ts-ignore
import type { ChapterSummaryId } from './chapter-summary-id';

export interface ChapterSummary {
    'id'?: ChapterSummaryId;
    'chapterId'?: ChapterId;
    'summaryType'?: ChapterSummarySummaryTypeEnum;
    'content'?: string;
    'keyPoints'?: Array<string>;
    'aiModel'?: string;
    'cached'?: boolean;
    'createTime'?: string;
}

export const ChapterSummarySummaryTypeEnum = {
    BRIEF: 'BRIEF',
    DETAILED: 'DETAILED',
    KEYPOINTS: 'KEYPOINTS'
} as const;

export type ChapterSummarySummaryTypeEnum = typeof ChapterSummarySummaryTypeEnum[keyof typeof ChapterSummarySummaryTypeEnum];