// web/src/api/generated/models/knowledge-point.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ChapterId } from './chapter-id';
// May contain unused imports in some cases
// @ts-ignore
import type { KnowledgePointId } from './knowledge-point-id';

export interface KnowledgePoint {
    'id'?: KnowledgePointId;
    'chapterId'?: ChapterId;
    'pointType'?: KnowledgePointPointTypeEnum;
    'name'?: string;
    'description'?: string;
    'position'?: number;
    'relatedChapterIds'?: Array<number>;
    'relatedPointIds'?: Array<number>;
    'createTime'?: string;
}

export const KnowledgePointPointTypeEnum = {
    CONCEPT: 'CONCEPT',
    TERM: 'TERM',
    FORMULA: 'FORMULA',
    PRINCIPLE: 'PRINCIPLE',
    METHOD: 'METHOD'
} as const;

export type KnowledgePointPointTypeEnum = typeof KnowledgePointPointTypeEnum[keyof typeof KnowledgePointPointTypeEnum];